Created navigation component. Also separated Admin and Home page.
