+++ package.json
@@ -15,6 +15,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.6",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
     "redux-immutable-state-invariant": "^2.1.0",
     "rxjs": "^7.5.2",
